separate web app into multiple files
