+++ README.md
@@ -62,12 +62,12 @@ is automatically generated by the AI after your first message.
 
 ## Web app
 
-A simple FastAPI-based web interface is available in `app.py`.
+A simple FastAPI-based web interface is available in `server.py`.
 Install the extra dependencies and run the server:
 
 ```bash
 pip install -r requirements.txt
-uvicorn app:app --host 0.0.0.0 --port 8000
+uvicorn server:app --host 0.0.0.0 --port 8000
 ```
 
 Open the displayed address in a browser on any device. The left sidebar lists
