+++ static/index.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset='utf-8'>
+  <title>GroqChat Web</title>
+  <style>
+    body{font-family:Arial,Helvetica,sans-serif;margin:0;display:flex;height:100vh;background:#1e1e1e;color:#eee}
+    #sidebar{width:250px;border-right:1px solid #444;overflow-y:auto;padding:10px;background:#2b2b2b;display:flex;flex-direction:column}
+    #tabButtons{display:flex;margin-bottom:10px}
+    .tab{flex:1;padding:5px;background:#333;border:1px solid #444;color:#eee;cursor:pointer;text-align:center}
+    .tab.active{background:#555}
+    #fileList{display:flex;flex-direction:column;gap:6px}
+    .chat-entry{cursor:pointer;color:#9cf;padding:2px}
+    .chat-entry:hover{background:#3a3a3a}
+    .chat-name{font-size:14px}
+    .chat-file{font-size:12px;color:#aaa;margin-left:4px}
+    #chat{flex:1;display:flex;flex-direction:column;height:100%}
+    #chatHeader{padding:10px;border-bottom:1px solid #444}
+    #chatName{font-size:18px;margin-bottom:2px}
+    #chatPath{font-size:12px;color:#aaa}
+    #summaryBox{margin-top:8px;font-size:12px;white-space:pre-wrap}
+    #messages{flex:1;overflow-y:auto;padding:10px;display:flex;flex-direction:column;gap:8px}
+    .message{padding:8px;border-radius:4px;max-width:80%;white-space:pre-wrap}
+    .user{background:#2f3b55;align-self:flex-end}
+    .assistant{background:#353535}
+    .system{background:#444;align-self:center}
+    .error{background:#552222;color:#ffbbbb;align-self:center}
+    #sysBox{margin-top:10px}
+    #sysPrompt{width:100%;background:#333;color:#eee;border:1px solid #555;margin-top:4px}
+    #input{display:flex;border-top:1px solid #444}
+    #input textarea{flex:1;padding:5px;background:#333;color:#eee;border:1px solid #444}
+    #input button{background:#444;color:#eee;border:1px solid #555;padding:5px 10px}
+  </style>
+</head>
+<body>
+  <div id='sidebar'>
+    <h3>Chats</h3>
+    <div id='tabButtons'></div>
+    <div id='fileList'></div>
+  </div>
+  <div id='chat'>
+    <div id='chatHeader'>
+      <div id='chatName'></div>
+      <div id='chatPath'></div>
+      <div id='summaryBox'></div>
+      <details id='sysBox'>
+        <summary>System Prompt</summary>
+        <textarea id='sysPrompt' rows='4'></textarea>
+        <button onclick='updateSystem()'>Save</button>
+      </details>
+    </div>
+    <div id='messages'></div>
+    <div id='input'>
+      <textarea id='msg' rows='3'></textarea>
+      <button onclick='sendMsg()'>Send</button>
+    </div>
+  </div>
+  <script src="app.js"></script>
+</body>
+</html>
